[23] Deprecate the ServerHelper and move the utilities to new utilities that are more appropriate for the server running.
