+++ README.adoc
@@ -83,9 +83,10 @@ try (final ModelControllerClient client = ModelControllerClient.Factory.create(I
 }
 ----
 
-== Server Helper
+== Server Utilities
 
-You can also use the `org.wildfly.plugin.core.ServerHelper` to interact with a running server.
+You can also use the `org.wildfly.plugin.tools.server.DomainManagement`,  `org.wildfly.plugin.tools.server.ServerManagement`
+and `org.wildfly.plugin.tools.server.StandaloneManagement` utilities to interact with a running server.
 
 [source,java]
 ----
@@ -93,14 +94,14 @@ final Path wildflyHome = Paths.get(System.getProperty("user.home"), "servers", "
 final Process process = Launcher.of(StandaloneCommandBuilder.of(wildflyHome)).launch();
 try (final ModelControllerClient client = ModelControllerClient.Factory.create(InetAddress.getLocalHost(), 9990)) {
     // Wait at the maximum 30 seconds for the server to start
-    ServerHelper.waitForStandalone(client, 30);
+    StandaloneManagement.waitForStandalone(client, 30);
     final Path deploymentPath = Paths.get(System.getProperty("user.home"), "projects", "myapp", "target", "myapp.war");
     final DeploymentManager deploymentManager = DeploymentManager.Factory.create(client);
     final Deployment deployment = Deployment.of(deploymentPath);
     deploymentManager.forceDeploy(deployment).assertSuccess();
 
     // Shutdown the standalone server
-    ServerHelper.shutdownStandalone(client);
+    StandaloneManagement.shutdownStandalone(client);
 } finally {
     process.destroy();
 }
