+++ src/main/java/org/wildfly/plugin/tools/server/DomainManagement.java
@@ -1,0 +1,223 @@
+/*
+ * Copyright The WildFly Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.wildfly.plugin.tools.server;
+
+import static org.jboss.as.controller.client.helpers.ClientConstants.CONTROLLER_PROCESS_STATE_STARTING;
+import static org.jboss.as.controller.client.helpers.ClientConstants.CONTROLLER_PROCESS_STATE_STOPPING;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+
+import org.jboss.as.controller.client.ModelControllerClient;
+import org.jboss.as.controller.client.helpers.Operations;
+import org.jboss.as.controller.client.helpers.domain.DomainClient;
+import org.jboss.as.controller.client.helpers.domain.ServerIdentity;
+import org.jboss.as.controller.client.helpers.domain.ServerStatus;
+import org.jboss.dmr.ModelNode;
+import org.jboss.logging.Logger;
+import org.wildfly.common.Assert;
+import org.wildfly.plugin.tools.OperationExecutionException;
+
+/**
+ * A utility for executing management operations on domain servers.
+ *
+ * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
+ */
+@SuppressWarnings("unused")
+public class DomainManagement extends ServerManagement {
+    private static final Logger LOGGER = Logger.getLogger(DomainManagement.class);
+
+    /**
+     * Waits the given amount of time in seconds for a managed domain to start. A domain is considered started when each
+     * of the servers in the domain are started unless the server is disabled.
+     *
+     * @param client         the client used to communicate with the server
+     * @param startupTimeout the time, in seconds, to wait for the server start
+     *
+     * @throws InterruptedException if interrupted while waiting for the server to start
+     * @throws RuntimeException     if the process has died
+     * @throws TimeoutException     if the timeout has been reached and the server is still not started
+     */
+    public static void waitForDomain(final ModelControllerClient client, final long startupTimeout)
+            throws InterruptedException, RuntimeException, TimeoutException {
+        waitForDomain(null, client, startupTimeout);
+    }
+
+    /**
+     * Waits the given amount of time in seconds for a managed domain to start. A domain is considered started when each
+     * of the servers in the domain are started unless the server is disabled.
+     * <p>
+     * If the {@code process} is not {@code null} and a timeout occurs the process will be
+     * {@linkplain Process#destroy() destroyed}.
+     * </p>
+     *
+     * @param process        the Java process can be {@code null} if no process is available
+     * @param client         the client used to communicate with the server
+     * @param startupTimeout the time, in seconds, to wait for the server start
+     *
+     * @throws InterruptedException if interrupted while waiting for the server to start
+     * @throws RuntimeException     if the process has died
+     * @throws TimeoutException     if the timeout has been reached and the server is still not started
+     */
+    public static void waitForDomain(final Process process, final ModelControllerClient client, final long startupTimeout)
+            throws InterruptedException, RuntimeException, TimeoutException {
+        Assert.checkNotNullParam("client", client);
+        long timeout = startupTimeout * 1000;
+        final long sleep = 100;
+        while (timeout > 0) {
+            long before = System.currentTimeMillis();
+            if (isDomainRunning(client)) {
+                break;
+            }
+            timeout -= (System.currentTimeMillis() - before);
+            if (process != null && !process.isAlive()) {
+                throw new RuntimeException(
+                        String.format("The process has unexpectedly exited with code %d", process.exitValue()));
+            }
+            TimeUnit.MILLISECONDS.sleep(sleep);
+            timeout -= sleep;
+        }
+        if (timeout <= 0) {
+            if (process != null) {
+                process.destroy();
+            }
+            throw new TimeoutException(String.format("The server did not start within %s seconds.", startupTimeout));
+        }
+    }
+
+    /**
+     * Checks to see if the domain is running. If the server is not in admin only mode each servers running state is
+     * checked. If any server is not in a started state the domain is not considered to be running.
+     *
+     * @param client the client used to communicate with the server
+     *
+     * @return {@code true} if the server is in a running state, otherwise {@code false}
+     */
+    public static boolean isDomainRunning(final ModelControllerClient client) {
+        return isDomainRunning(client, false);
+    }
+
+    /**
+     * Shuts down a managed domain container. The servers are first stopped, then the host controller is shutdown.
+     *
+     * @param client the client used to communicate with the server
+     *
+     * @throws IOException                 if an error occurs communicating with the server
+     * @throws OperationExecutionException if the operation used to shutdown the managed domain failed
+     */
+    public static void shutdownDomain(final ModelControllerClient client) throws IOException, OperationExecutionException {
+        shutdownDomain(client, 0);
+    }
+
+    /**
+     * Shuts down a managed domain container. The servers are first stopped, then the host controller is shutdown.
+     *
+     * @param client  the client used to communicate with the server
+     * @param timeout the graceful shutdown timeout, a value of {@code -1} will wait indefinitely and a value of
+     *                    {@code 0} will not attempt a graceful shutdown
+     *
+     * @throws IOException                 if an error occurs communicating with the server
+     * @throws OperationExecutionException if the operation used to shutdown the managed domain failed
+     */
+    public static void shutdownDomain(final ModelControllerClient client, final int timeout)
+            throws IOException, OperationExecutionException {
+        // Note the following two operations used to shutdown a domain don't seem to work well in a composite operation.
+        // The operation occasionally sees a java.util.concurrent.CancellationException because the operation client
+        // is likely closed before the AsyncFuture.get() is complete. Using a non-composite operation doesn't seem to
+        // have this issue.
+
+        // First shutdown the servers
+        final ModelNode stopServersOp = Operations.createOperation("stop-servers");
+        stopServersOp.get("blocking").set(true);
+        stopServersOp.get("timeout").set(timeout);
+        ModelNode response = client.execute(stopServersOp);
+        if (!Operations.isSuccessfulOutcome(response)) {
+            throw new OperationExecutionException("Failed to stop servers.", stopServersOp, response);
+        }
+
+        // Now shutdown the host
+        final ModelNode address = determineHostAddress(client);
+        final ModelNode shutdownOp = Operations.createOperation("shutdown", address);
+        response = client.execute(shutdownOp);
+        if (Operations.isSuccessfulOutcome(response)) {
+            // Wait until the process has died
+            while (true) {
+                if (isDomainRunning(client, true)) {
+                    Thread.onSpinWait();
+                } else {
+                    break;
+                }
+            }
+        } else {
+            throw new OperationExecutionException("Failed to shutdown host.", shutdownOp, response);
+        }
+    }
+
+    /**
+     * Determines the address for the host being used.
+     *
+     * @param client the client used to communicate with the server
+     *
+     * @return the address of the host
+     *
+     * @throws IOException                 if an error occurs communicating with the server
+     * @throws OperationExecutionException if the operation used to determine the host name fails
+     */
+    public static ModelNode determineHostAddress(final ModelControllerClient client)
+            throws IOException, OperationExecutionException {
+        final ModelNode op = Operations.createReadAttributeOperation(EMPTY_ADDRESS, "local-host-name");
+        ModelNode response = client.execute(op);
+        if (Operations.isSuccessfulOutcome(response)) {
+            return Operations.createAddress("host", Operations.readResult(response).asString());
+        }
+        throw new OperationExecutionException(op, response);
+    }
+
+    private static boolean isDomainRunning(final ModelControllerClient client, boolean shutdown) {
+        final DomainClient domainClient = (client instanceof DomainClient ? (DomainClient) client
+                : DomainClient.Factory.create(client));
+        try {
+            // Check for admin-only
+            final ModelNode hostAddress = determineHostAddress(domainClient);
+            final Operations.CompositeOperationBuilder builder = Operations.CompositeOperationBuilder.create()
+                    .addStep(Operations.createReadAttributeOperation(hostAddress, "running-mode"))
+                    .addStep(Operations.createReadAttributeOperation(hostAddress, "host-state"));
+            ModelNode response = domainClient.execute(builder.build());
+            if (Operations.isSuccessfulOutcome(response)) {
+                response = Operations.readResult(response);
+                if ("ADMIN_ONLY".equals(Operations.readResult(response.get("step-1")).asString())) {
+                    if (Operations.isSuccessfulOutcome(response.get("step-2"))) {
+                        final String state = Operations.readResult(response).asString();
+                        return !CONTROLLER_PROCESS_STATE_STARTING.equals(state)
+                                && !CONTROLLER_PROCESS_STATE_STOPPING.equals(state);
+                    }
+                }
+            }
+            final Map<ServerIdentity, ServerStatus> servers = new HashMap<>();
+            final Map<ServerIdentity, ServerStatus> statuses = domainClient.getServerStatuses();
+            for (ServerIdentity id : statuses.keySet()) {
+                final ServerStatus status = statuses.get(id);
+                switch (status) {
+                    case DISABLED:
+                    case STARTED: {
+                        servers.put(id, status);
+                        break;
+                    }
+                }
+            }
+            if (shutdown) {
+                return statuses.isEmpty();
+            }
+            return statuses.size() == servers.size();
+        } catch (Exception e) {
+            LOGGER.trace("Interrupted determining if domain is running", e);
+        }
+        return false;
+    }
+}
