+++ src/test/java/org/wildfly/plugin/tools/DomainDeploymentManagerIT.java
@@ -23,6 +23,8 @@ import org.junit.jupiter.api.Test;
 import org.wildfly.core.launcher.DomainCommandBuilder;
 import org.wildfly.core.launcher.Launcher;
 import org.wildfly.core.launcher.ProcessHelper;
+import org.wildfly.plugin.tools.server.DomainManagement;
+import org.wildfly.plugin.tools.server.StandaloneManagement;
 
 /**
  * @author <a href="mailto:jperkins@redhat.com">James R. Perkins</a>
@@ -64,8 +66,8 @@ public class DomainDeploymentManagerIT extends AbstractDeploymentManagerTest {
         boolean ok = false;
         try {
             client = DomainClient.Factory.create(Environment.createClient());
-            if (ServerHelper.isDomainRunning(client) || ServerHelper.isStandaloneRunning(client)) {
-                Assertions.fail("A WildFly server is already running: " + ServerHelper.getContainerDescription(client));
+            if (DomainManagement.isDomainRunning(client) || StandaloneManagement.isStandaloneRunning(client)) {
+                Assertions.fail("A WildFly server is already running: " + ContainerDescription.lookup(client));
             }
             final DomainCommandBuilder commandBuilder = DomainCommandBuilder.of(Environment.WILDFLY_HOME);
             if (IS_MODULAR_JDK) {
@@ -73,7 +75,7 @@ public class DomainDeploymentManagerIT extends AbstractDeploymentManagerTest {
             }
             process = Launcher.of(commandBuilder).launch();
             consoleConsomer = ConsoleConsumer.start(process, System.out);
-            ServerHelper.waitForDomain(client, Environment.TIMEOUT);
+            DomainManagement.waitForDomain(client, Environment.TIMEOUT);
             ok = true;
         } finally {
             if (!ok) {
@@ -95,7 +97,7 @@ public class DomainDeploymentManagerIT extends AbstractDeploymentManagerTest {
     public static void shutdown() throws Exception {
         try {
             if (client != null) {
-                ServerHelper.shutdownDomain(client);
+                DomainManagement.shutdownDomain(client);
                 safeClose(client);
             }
         } finally {
@@ -184,7 +186,7 @@ public class DomainDeploymentManagerIT extends AbstractDeploymentManagerTest {
 
     @Override
     protected ModelNode createDeploymentResourceAddress(final String deploymentName) throws IOException {
-        return ServerHelper.determineHostAddress(getClient())
+        return DomainManagement.determineHostAddress(getClient())
                 .add(ClientConstants.SERVER, "server-one")
                 .add(ClientConstants.DEPLOYMENT, deploymentName);
     }
